Move configuration merging to @autorest/configuration
