+++ packages/extensions/core/src/app.ts
@@ -58,7 +58,7 @@ import { Help } from "./help";
 import { CreateConfiguration, isLegacy } from "./legacyCli";
 import { Artifact } from "./lib/artifact";
 import { AutoRest, IsOpenApiDocument, Shutdown } from "./lib/autorest-core";
-import { mergeConfigurations } from "./lib/configuration";
+import { mergeConfigurations } from "@autorest/configuration";
 import { Exception } from "@autorest/common";
 import { Channel, Message } from "./lib/message";
 import { homedir } from "os";
