+++ packages/extensions/core/src/lib/configuration/index.ts
@@ -1,6 +1,4 @@
 export * from "./configuration-loader";
-export * from "./configuration-merging";
-export * from "./autorest-configuration";
 export * from "./autorest-context";
 export * from "./message-emitter";
 export * from "./configuration-file-resolver";
