+++ packages/libs/configuration/package.json
@@ -35,5 +35,10 @@
     "jest": "^26.6.3",
     "rimraf": "^3.0.2",
     "typescript": "~4.1.3"
+  },
+  "dependencies": {
+    "@autorest/common": "~1.0.0",
+    "@azure-tools/datastore": "~4.1.271",
+    "@azure-tools/uri": "~3.0.0"
   }
 }
