+++ packages/libs/configuration/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./autorest-configuration";
 export * from "./autorest-raw-configuration";
+export * from "./configuration-merging";
 export * from "./directive";
 export * from "./utils";
